\encoding{latin1}
\name{convert}
\alias{convert}

\title{Conversion from one unit to another one}

\description{A generic function for unit conversion}

\usage{
convert(x, ...)
}

\arguments{
  \item{x}{an object whose unit should be converted}
  \item{\dots}{further arguments}
}


\value{
  The same object as 'x', but with unit converted

}

\author{
Philippe Grosjean \email{phgrosjean@sciviews.org}
}


\examples{
# TODO...
}

\keyword{utilities}
