\name{seacarb-package}
\alias{seacarb-package}
\alias{seacarb3}
\alias{seacarb}
\docType{package}
\title{
Calculates parameters of the seawater carbonate system (additions and changes proposed to the original seacarb)
}
\description{
Calculates parameters of the seawater carbonate system
}
\details{
  \tabular{ll}{
    Package: \tab seacarb3\cr
    Type: \tab Package\cr
    Version: \tab 3.0\cr
    Date: \tab 2009-??-??\cr
    License: \tab GNU Public License 2 or above\cr
  }

}
\author{

  Heloise Lavigne, Aurelien Proye, Philippe Grosjean and Jean-Pierre Gattuso (maintainer).

  Portions of code and/or corrections were contributed by Jean-Marie Epitalon, Bernard Gentili, Jim Orr and Karline Soetaert
  
}
\references{

}

\examples{
\dontrun{
## show examples (see respective help pages for details)
example(bjerrum)
example(K1)

## open the directory with documents
browseURL(paste(system.file(package="seacarb3"), "/data", sep=""))


}
}

\keyword{ package }
